% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aire-package.R
\docType{package}
\name{aire-package}
\alias{aire}
\alias{aire-package}
\title{Grab data on air pollution in Andalusia}
\description{
This package can download and retrieve daily data on air quality in Andalusia (Spain),
from any of the stations maintained by 'Consejeria de Medio Ambiente' (http://bit.ly/airedata).
The data are formatted so that they can be directly analysed with the 'openair' package. Note these data
are collected automatically, and not validated. Check our Consejeria de Medio Ambiente website for monthly
and annual reports based on validated data.
Currently, data are available from 2005 onwards.
}
\keyword{package}

