% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get daily data for given dates and province(s)}
\usage{
get_daily_data(fecha, provincia)
}
\arguments{
\item{fecha}{Character vector with date in formart "yyyy-mm-dd" (see examples).}

\item{provincia}{Character. Two first letters of province name.}
}
\value{
List of characters with http addresses.
}
\description{
Get daily data for given dates and province(s)
}
\examples{

get_daily_data("2015-01-28", "gr")

get_daily_data("2015-01-03", "MA")

get_daily_data("2015-01-03", "Se")

get_daily_data(c("2015-01-28", "2015-01-29"), c("se", "gr"))

}
\author{
F. Rodriguez-Sanchez, V. Luque
}

