% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se2015.R
\docType{data}
\name{se2015}
\alias{se2015}
\title{Air quality data for Sevilla province, year 2015.}
\format{A dataframe with 606911 rows and 11 variables:
\describe{
  \item{Provincia}{Province}
  \item{Municipio}{Council where station is located}
  \item{site}{Station name}
  \item{Direccion}{Address}
  \item{date}{Date and time of the measurement, in format "yyyy-mm-dd hh:mm:ss"}
  \item{SO2}{SO2 concentration (in micrograms per cubic meter)}
  \item{PART}{Particles concentration (in micrograms per cubic meter)}
  \item{NO2}{NO2 concentration (in micrograms per cubic meter)}
  \item{CO}{CO concentration (in micrograms per cubic meter)}
  \item{O3}{O3 concentration (in micrograms per cubic meter)}
  \item{SH2}{SH2 concentration (in micrograms per cubic meter)}
}}
\source{
Consejeria de Medio Ambiente, Junta de Andalucia (\url{http://bit.ly/airedata})
}
\description{
Air quality data for all stations in Sevilla province for year 2015.
}

