% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Measurement stations info}
\format{A SpatialPointsDataFrame with 99 rows and 5 variables:
\describe{
  \item{X}{Station number}
  \item{ESTACION}{Station name}
  \item{MUNICIPIO}{Council}
  \item{PROVINCIA}{Province}
  \item{site}{Station name (used for compatibility with package openair)}
}}
\source{
Consejeria de Medio Ambiente, Junta de Andalucia (\url{http://www.juntadeandalucia.es/medioambiente/portal_web/web/temas_ambientales/atmosfera/calidad_aire/RVCCAA_2015.pdf})
}
\description{
Coordinates and information of all measurement stations in Andalusia.
}

