% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_finder.R
\name{url_finder}
\alias{url_finder}
\title{Conseguir rutas URL}
\usage{
url_finder(mes, dia)
}
\arguments{
\item{mes}{Dame las tres primeras letras del mes.}

\item{dia}{Dia juliano.}
}
\description{
Conseguir rutas URL
}
\author{
Paco
}

