% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{list2df
Toma una lista de lecturas de par치metros de calidad del aire
para un dia determinado para cada estaci칩n provincial y
lecturas de cas}
\usage{
list2df(list.est)
}
\arguments{
\item{list.est}{}
}
\value{
data.frame
}
\description{
list2df
Toma una lista de lecturas de par치metros de calidad del aire
para un dia determinado para cada estaci칩n provincial y
lecturas de cas
}
\examples{
 list.dia<-  html2list("http://www.juntadeandalucia.es/medioambiente/atmosfera/informes_siva/ene15/nse150129.htm")
 df.dia  <- list2df(list.dia)

}

